+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](<https://semver.org/spec/v2.0.
 
 ### Added
 
+- Setting overrides to Animations Editor.
 - `action` trigger. Activates animations whenever an action is posted, like Recognize Spell or Change Shape.
 - Animation History menu. Stores animations played in the current session. Removed on refresh.
 - Delay Animations settings. ([#111](<https://github.com/MrVauxs/pf2e-graphics/issues/111>))